GETSIMN command inclusion
