+++ ahnlich/dsl/src/lib.rs
@@ -1,5 +1,8 @@
+mod algorithm;
 pub mod db;
 pub mod error;
 mod metadata;
 mod parser;
 mod predicate;
+#[cfg(test)]
+mod tests;
